requirements.txt
streamlit
yt-dlp
openai-whisper
ffmpeg-python
librosa
numpy